Task_1.2: Вынесение метода класса MenuBarGenerator в MainApplicationFrame Изменение вызова события закрытия окна
